backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pillow==10.3.0
python-multipart==0.0.9
bcrypt==4.0.1
python-jose~=3.5.0
ultralytics~=8.3.159
pydantic~=2.11.7
requests~=2.32.4
bs4~=0.0.2
beautifulsoup4~=4.13.4